Rename tools.actors to tools.concurrency
Make name more general to account for additional concurrency models.
