+++ src/som/interpreter/SomLanguage.java
@@ -22,10 +22,10 @@ import som.compiler.MixinDefinition;
 import som.vm.NotYetImplementedException;
 import som.vm.constants.Nil;
 import som.vmobjects.SClass;
-import tools.actors.Tags.ChannelRead;
-import tools.actors.Tags.ChannelWrite;
-import tools.actors.Tags.EventualMessageSend;
-import tools.actors.Tags.ExpressionBreakpoint;
+import tools.concurrency.Tags.ChannelRead;
+import tools.concurrency.Tags.ChannelWrite;
+import tools.concurrency.Tags.EventualMessageSend;
+import tools.concurrency.Tags.ExpressionBreakpoint;
 import tools.debugger.Tags.ArgumentTag;
 import tools.debugger.Tags.CommentTag;
 import tools.debugger.Tags.DelimiterClosingTag;
