+++ src/som/interpreter/actors/EventualSendNode.java
@@ -31,8 +31,8 @@ import som.vm.constants.Nil;
 import som.vmobjects.SSymbol;
 import tools.SourceCoordinate;
 import tools.SourceCoordinate.FullSourceCoordinate;
-import tools.actors.Tags.EventualMessageSend;
-import tools.actors.Tags.ExpressionBreakpoint;
+import tools.concurrency.Tags.EventualMessageSend;
+import tools.concurrency.Tags.ExpressionBreakpoint;
 import tools.debugger.nodes.AbstractBreakpointNode;
 import tools.debugger.nodes.BreakpointNodeGen;
 import tools.debugger.nodes.DisabledBreakpointNode;
