+++ src/som/primitives/processes/ChannelPrimitives.java
@@ -38,9 +38,9 @@ import som.vmobjects.SObjectWithClass;
 import som.vmobjects.SSymbol;
 import tools.SourceCoordinate;
 import tools.SourceCoordinate.FullSourceCoordinate;
-import tools.actors.Tags.ChannelRead;
-import tools.actors.Tags.ChannelWrite;
-import tools.actors.Tags.ExpressionBreakpoint;
+import tools.concurrency.Tags.ChannelRead;
+import tools.concurrency.Tags.ChannelWrite;
+import tools.concurrency.Tags.ExpressionBreakpoint;
 import tools.debugger.nodes.AbstractBreakpointNode;
 import tools.debugger.nodes.BreakpointNodeGen;
 import tools.debugger.nodes.DisabledBreakpointNode;
