+++ src/tools/concurrency/Tags.java
@@ -1,4 +1,4 @@
-package tools.actors;
+package tools.concurrency;
 
 
 public abstract class Tags {
