+++ src/tools/debugger/FrontendConnector.java
@@ -26,8 +26,8 @@ import som.vm.VmSettings;
 import som.vmobjects.SSymbol;
 import tools.SourceCoordinate;
 import tools.SourceCoordinate.TaggedSourceCoordinate;
+import tools.concurrency.ActorExecutionTrace;
 import tools.Tagging;
-import tools.actors.ActorExecutionTrace;
 import tools.debugger.frontend.Suspension;
 import tools.debugger.message.Message;
 import tools.debugger.message.Message.OutgoingMessage;
